commit agradecimiento personalizado funcional

+++ carrito.js
@@ -50,6 +50,18 @@ function logout(){
     mostrarUsuarioLogueado();
 }
 
+function usuarioLogueado(){
+    let userLog = JSON.parse(localStorage.getItem("USER")) || [];
+
+    if (userLog.length != 0) { //hay usuario logueado
+        const textoUser = document.getElementById("user");
+        return ", " + userLog.apeYnom;
+    } else {
+        return "";
+    }
+    return ;
+}
+
 
 async function login() {
     let usuario;
@@ -66,39 +78,44 @@ async function login() {
             ]
         }
     })
-    // console.log(usuario);
+    console.log(usuario);
 
-    const {
-        value: password
-    } = await Swal.fire({
-        title: 'Ingresa tu password',
-        input: 'password',
-        inputLabel: 'Password',
-        inputPlaceholder: 'Enter your password',
-        inputAttributes: {
-            maxlength: 10,
-            autocapitalize: 'off',
-            autocorrect: 'off'
-        }
-    })
-
-    let exist = arrayUsuarios.find((element) => element.dni == usuario & element.pass == password);
-    let indiceExist = arrayUsuarios.indexOf(exist);
-
-    if (indiceExist != -1) {
-        const textoLogin = document.getElementById("divLogin");
-        textoLogin.innerHTML = `<a href="#" id="login" class="btn btn-ligth" width="20" height="20" onclick="logout();">Logout</a>`;
-        localStorage.setItem('USER', JSON.stringify(arrayUsuarios[indiceExist]));
-        mostrarUsuarioLogueado();
-    } else { //no existe usuario o contraseña erronea
-        Swal.fire({
-            icon: 'error',
-            title: 'Oops...',
-            text: 'El usuario no existe o la contraseña es incorrecta!',
-            //footer: '<a href="">Why do I have this issue?</a>'
+    if (usuario != undefined) {
+        const {
+            value: password
+        } = await Swal.fire({
+            title: 'Ingresa tu password',
+            input: 'password',
+            inputLabel: 'Password',
+            inputPlaceholder: 'Enter your password',
+            inputAttributes: {
+                maxlength: 10,
+                autocapitalize: 'off',
+                autocorrect: 'off'
+            }
         })
+
+        let exist = arrayUsuarios.find((element) => element.dni == usuario & element.pass == password);
+        let indiceExist = arrayUsuarios.indexOf(exist);
+    
+        if (indiceExist != -1) {
+            const textoLogin = document.getElementById("divLogin");
+            textoLogin.innerHTML = `<a href="#" id="login" class="btn btn-ligth" width="20" height="20" onclick="logout();">Logout</a>`;
+            localStorage.setItem('USER', JSON.stringify(arrayUsuarios[indiceExist]));
+            mostrarUsuarioLogueado();
+        } else { //no existe usuario o contraseña erronea
+            Swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'El usuario no existe o la contraseña es incorrecta!',
+                //footer: '<a href="">Why do I have this issue?</a>'
+            })
+        }
     }
 }
+
+
+
 
 
 function finalizarCompra (){
@@ -416,10 +433,11 @@ botonFinalizar.addEventListener("click", ()=> {
 
     finalizarCompra();
 
+
     Swal.fire({
         position: 'center',
         icon: 'success',
-        title: 'Pedido Realizado. <br>Muchas gracias!!',
+        title: 'Pedido Realizado. <br>Muchas gracias' + usuarioLogueado() + '!!',
         showConfirmButton: false,
         timer: 3500
     })
